+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * fix do macro inside syntax-rules [#81](https://github.com/jcubic/lips/issues/81)
 * fix `string-copy` and `string-set!`
 * fix `error` function to work like R7RS spec
+* fix edge case in `syntax-rules` found when executing SRFI-26 cut macro
 
 ## 1.0.0-beta.7
 ### Breaking